agent-container/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
strands-agents
boto3>=1.35.0
strands-agents-tools